Обновление: MediaInfo git-v0.7.98-2-g3436139b; Обновление:ZenLib git-v0.4.36.
git-svn-id: svn://svn.code.sf.net/p/mpcbe/code/trunk@2807 7426f7ed-705a-47fc-8cda-d59144bb0dce

+++ src/ExtLib/ZenLib/ZenLib/int128u.cpp
@@ -97,7 +97,7 @@ uint128::uint128 (const char * sz) throw ()
             ++i;
         };
     };
-    const unsigned int len = strlen(sz);
+    const size_t len = strlen(sz);
     for (; i < len; ++i) {
         unsigned int n = 0;
         if (sz [i] >= '0' && sz [i] <= (('0' + (int) radix) < '9'?('0' + (int) radix):'9')) //if (sz [i] >= '0' && sz [i] <= (('0' + (int) radix) <? '9'))
